v2/requirements.txt
flask
flask-cors
pymupdf
openai
Pillow
requests
gunicorn
psycopg2-binary
